use icon component instead of svg element
