+++ resources/js/Ab/DeleteButton.vue
@@ -1,12 +1,15 @@
 <template>
     <button class="font-medium text-red-600 hover:text-red-900 focus:outline-none">
         <slot>
-            <svg class="h-6 w-6" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M10 14l2-2m0 0l2-2m-2 2l-2-2m2 2l2 2m7-2a9 9 0 11-18 0 9 9 0 0118 0z"/></svg>
+            <ab-icon name="x-circle" size="md" />
         </slot>
     </button>
 </template>
 
 <script>
+    import AbIcon from './Icon'
+
     export default {
+        components: {AbIcon},
     }
 </script>
