+++ resources/js/Ab/DirectionButton.vue
@@ -1,14 +1,19 @@
 <template>
     <button class="font-medium text-blue-600 hover:text-blue-900 focus:outline-none">
         <slot>
-            <svg v-if="direction == 'up'" class="h-6 w-6" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 11l3-3m0 0l3 3m-3-3v8m0-13a9 9 0 110 18 9 9 0 010-18z"/></svg>
-            <svg v-else class="h-6 w-6" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor"><path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M15 13l-3 3m0 0l-3-3m3 3V8m0 13a9 9 0 110-18 9 9 0 010 18z"/></svg>
+            <ab-icon v-if="direction == 'up'" name="arrow-circle-up" size="md" />
+            <ab-icon v-else name="arrow-circle-down" size="md" />
         </slot>
     </button>
 </template>
 
 <script>
+    import AbIcon from './Icon'
+
     export default {
+        components: {
+            AbIcon,
+        },
         props: {
             direction: {
                 type: String,
